Insert source folder and rename agents module
