+++ agentMet4FoF/dashboard/Dashboard.py
@@ -9,10 +9,10 @@ import plotly.graph_objs as go
 from plotly import tools
 
 import networkx as nx
-import AgentMET4FOF as agentmet4fof_module
+import agentMet4FoF.agents as agentmet4fof_module
 
-import dashboard.LayoutHelper as LayoutHelper
-from dashboard.LayoutHelper import create_nodes_cytoscape, create_edges_cytoscape, create_monitor_graph
+import agentMet4FoF.dashboard.LayoutHelper as LayoutHelper
+from agentMet4FoF.dashboard.LayoutHelper import create_nodes_cytoscape, create_edges_cytoscape, create_monitor_graph
 
 external_stylesheets = ['https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/css/materialize.min.css', 'https://fonts.googleapis.com/icon?family=Material+Icons']
 external_scripts = ['https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/js/materialize.min.js']
