+++ tests/test_send_plot.py
@@ -1,4 +1,4 @@
-from AgentMET4FOF import AgentMET4FOF, AgentNetwork, MonitorAgent
+from agentMet4FoF.agents import AgentMET4FOF, AgentNetwork, MonitorAgent
 import matplotlib.pyplot as plt
 import numpy as np
 import time
