+++ tests/test_zema_emc_ml.py
@@ -1,4 +1,4 @@
-from AgentMET4FOF import AgentNetwork, MonitorAgent, DataStreamAgent
+from agentMet4FoF.agents import AgentNetwork, MonitorAgent, DataStreamAgent
 
 
 from examples.ZEMA_EMC.zema_agents import TrainTestSplitAgent, FFT_BFCAgent, Pearson_FeatureSelectionAgent, LDA_Agent, EvaluatorAgent
