+++ tutorials/tutorial_1_generator_agent.py
@@ -1,5 +1,5 @@
-from AgentMET4FOF import AgentMET4FOF, AgentNetwork, MonitorAgent
-from DataStreamMET4FOF import SineGenerator
+from agentMet4FoF.agents import AgentMET4FOF, AgentNetwork, MonitorAgent
+from agentMet4FoF.streams import SineGenerator
 
 #Here we define a new agent SineGeneratorAgent, and override the functions : init_parameters & agent_loop
 #init_parameters() is used to setup the data stream and necessary parameters
